Add the ability to use an alternative API server to perform requests against the cluster.
This has some far-reaching consequences.

Because we're now using the identity of the caller of the HTTP request to
perform the calls into the cluster, we need to make sure that the user is
properly authenticated and that we don't escalate the permissions allowed
for the user by using the service account of the OAuth service.

This in turn requires us to reintroduce to somehow remember the identity
of the user during the OAuth flow. This is now done such that each OAuth
flow is assigned a unique key that is stored in the OAuth state and in the
session of the OAuth service (the session id is stored as a cookie). The
bearer token is kept in the session in-memory on the server side. Upon
completing the OAuth flow the key from the OAuth state is used to look up
the bearer token in the session. The OAuth flow fails if the session
doesn't contain the bearer token under the key.

This way we can make sure that the user that started the OAuth flow is the
same as the one who finishes it (by relying on the session).